+++ src/basilisp_blender/__init__.py
@@ -5,6 +5,7 @@ import logging
 
 from basilisp import main as basilisp
 from basilisp.lang import compiler
+from basilisp.lang.keyword import Keyword
 from basilisp.lang.util import munge
 
 COMPILER_OPTS = compiler.compiler_opts()
@@ -12,7 +13,6 @@ basilisp.init(COMPILER_OPTS)
 
 LOGGER = logging.getLogger("basilisp-blender")
 LOGGER.addHandler(logging.StreamHandler())
-
 
 def log_level_set(level, filepath=None):
     """Sets the logger in the `LOGGER` global variable to the
@@ -31,6 +31,13 @@ def log_level_set(level, filepath=None):
 # log_level_set(logging.DEBUG, "basilisp-blender.log")
 
 def control_panel_create():
-    """Initialises and displays the nREPL server UI control panel."""
+    """Initialises and displays the nREPL server UI control panel. It
+    returns a function to destroy the panel and settings, and stop the
+    server if it is running.
+
+    """
     ctrl_panel_mod = importlib.import_module(munge("basilisp-blender.control-panel"))
-    ctrl_panel_mod.nrepl_control_panel_create__BANG__()
+    panel = ctrl_panel_mod.nrepl_control_panel_create__BANG__()
+    return panel[Keyword("destroy!")]
+
+